radio: add form example and submission logic
