+++ packages/radio/examples/basic.example.tsx
@@ -10,10 +10,10 @@ function Example() {
       <h2>HTML input</h2>
       <h3 id="fave">Which number is your favorite?</h3>
       <fieldset aria-labelledby="fave" style={{ border: 0, padding: 0 }}>
-        <input type="radio" id="fave-one" name="faves-1" value="one" />
+        <input disabled type="radio" id="fave-one" name="faves-1" value="one" />
         <label htmlFor="fave-one">One</label>
 
-        <input disabled type="radio" id="fave-two" name="faves-1" value="two" />
+        <input type="radio" id="fave-two" name="faves-1" value="two" />
         <label htmlFor="fave-two">Two</label>
 
         <input type="radio" id="fave-three" name="faves-1" value="three" />
@@ -23,10 +23,10 @@ function Example() {
       <h2>@reach/radio</h2>
       <h3 id="second-fave">Which number is your second favorite?</h3>
       <RadioGroup aria-labelledby="second-fave" name="faves-2">
-        <Radio value="one">One</Radio>
-        <Radio disabled value="two">
-          Two
+        <Radio disabled value="one">
+          One
         </Radio>
+        <Radio value="two">Two</Radio>
         <Radio value="three">Three</Radio>
       </RadioGroup>
     </div>
